Splitting app.py into smaller pieces. Cleaning up logging and old code
