+++ backend/requirements.txt
@@ -1,6 +1,6 @@
 fastapi==0.104.0
 uvicorn==0.23.2
-pydantic==2.4.2
+pydantic==2.11.4
 python-dotenv==1.0.0
 azure-cosmos==4.5.1
 azure-functions==1.17.0
