+++ backend/src/routes/__init__.py
@@ -1,0 +1,4 @@
+"""
+Routes package for the Azure Chat application.
+This package contains all the API route modules.
+"""
